Moved AttackFlags to Rhisis.World/Game/Common
